nass_quickstats_remote_provider/boson/requirements.txt
geodesic-api
boson-sdk[all]
geopandas
pyarrow
cachetools
